// src/app/core/api/models/expense-item-detail.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AccountDetail } from '../models/account-detail';
export interface ExpenseItemDetail {
  amount?: number;
  description?: string;
  expenseAccount?: AccountDetail;
  expenseDate?: string;
  id?: string;
  itemName?: string;
  remove?: boolean;
  status?: 'SUBMITTED' | 'PAID' | 'CLOSED';
}